refactor(tests): remove redundant tests, simplify test structure
