--- malimo.Tests/data/seed/ImagesInSubdirectory.md
@@ -1,9 +1,0 @@
-# Test Input
-
-This file contains some markdown to test the program.
-
-## Plant
-
-![[noun-plant-5382590.png]]
-
-Plant by Arkinasi from <a href="https://thenounproject.com/browse/icons/term/plant/" target="_blank" title="Plant Icons">Noun Project</a>
