--- malimo.Tests/data/seed/Testfile.md
@@ -1,15 +1,0 @@
-# Test Input
-
-This file contains some markdown to test the program.
-
-## Starship
-
-![[noun-starship-3799189.png]]
-
-starship by Made from <a href="https://thenounproject.com/browse/icons/term/starship/" target="_blank" title="starship Icons">Noun Project</a>
-
-## Island
-
-![[noun-island-1479438.png]]
-
-island by darwis from <a href="https://thenounproject.com/browse/icons/term/island/" target="_blank" title="island Icons">Noun Project</a>
