+++ malimo.Tests/malimo.Tests.csproj
@@ -23,7 +23,6 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\malimo.TestDirectoryHelper\malimo.TestDirectoryHelper.csproj" />
         <ProjectReference Include="..\malimo\malimo.csproj" />
     </ItemGroup>
